add routeing and pages:about,frontpage,ontology
