+++ frontend/package.json
@@ -16,6 +16,7 @@
     "@types/react": "^17.0.0",
     "@types/react-dom": "^17.0.0",
     "@types/react-redux": "^7.1.16",
+    "@types/react-router-dom": "^5.1.7",
     "d3": "^6.5.0",
     "eslint-config-airbnb-typescript": "^12.3.1",
     "eslint-config-react": "^1.1.7",
@@ -28,6 +29,7 @@
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-redux": "^7.2.2",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.2",
     "redux": "^4.0.5",
     "typescript": "^4.1.3",
