+++ frontend/src/pages/About.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const About = () => {
+  const test = { test1: String };
+  console.log(test);
+
+  return <div>test1</div>;
+};
+
+export default About;
